+++ .travis/check-posix-app-pty
@@ -47,7 +47,7 @@ at_exit() {
 }
 
 build() {
-    make -f examples/Makefile-sim
+    make -f examples/Makefile-simulation
     make -f src/posix/Makefile-posix PLATFORM_NETIF=1 PLATFORM_UDP=1 UDP_FORWARD=0
 }
