+++ etc/docker/ot_sim/start_sim
@@ -1,2 +1,2 @@
 #!/bin/bash
-docker run --rm -d --name otsim openthread/sim tail -F /dev/null
+docker run --rm -d --name otsim openthread/simulation tail -F /dev/null
